requirements.txt
streamlit>=1.28.0
aiohttp>=3.8.0
google-generativeai>=0.3.0